002-rag-retriever/requirements.txt
langchain
langchain-openai
langchain-community
streamlit
python-dotenv
langsmith
langchain-chroma